# specs/001-drive-narrator-voice-first/openapi.yaml
openapi: 3.0.3
info:
  title: Drive Narrator Voice-First Companion API
  description: |
    API for the Drive Narrator voice-first companion app that provides location-aware storytelling 
    and POI narration during drives. Designed for voice-first interaction with minimal visual dependency.
  version: 1.0.0
  contact:
    name: Drive Narrator Team
    email: support@drivenarrator.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.drivenarrator.com/v1
    description: Production server
  - url: https://staging-api.drivenarrator.com/v1
    description: Staging server
  - url: http://localhost:41234/api
    description: Development server
paths:
...
  /drives:
    post:
      summary: Start new drive session
      description: Create a new drive session for voice-first narration
      operationId: startDrive
      tags:
        - Drives
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriveStartRequest'
      responses:
        '201':
          description: Drive session started successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveSession'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /drives/{driveId}:
    get:
      summary: Get drive session
      description: Retrieve drive session information and status
      operationId: getDrive
      tags:
        - Drives
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Drive session ID
      responses:
        '200':
          description: Drive session retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveSession'
        '404':
          description: Drive session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: Update drive session
      description: Update drive session status or settings
      operationId: updateDrive
      tags:
        - Drives
      parameters:
        - name: driveId
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Drive session ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DriveUpdateRequest'
      responses:
        '200':
          description: Drive session updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DriveSession'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Drive session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Voice Conversation
  /conversation:
    post:
      summary: Process voice conversation
      description: Process voice input and return conversational response
      operationId: processConversation
      tags:
        - Conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationRequest'
      responses:
        '200':
          description: Conversation processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # POI Management
...
components:
  schemas:
...
    DriveSession:
      type: object
      required:
        - driveId
        - profileId
        - startTime
        - status
        - createdAt
        - updatedAt
      properties:
        driveId:
          type: string
          format: uuid
          description: Unique drive session identifier
        profileId:
          type: string
          format: uuid
          description: Associated user profile ID
        startTime:
          type: string
          format: date-time
          description: Drive start timestamp
        endTime:
          type: string
          format: date-time
          description: Drive end timestamp
        status:
          type: string
          enum: [active, paused, completed, archived, cancelled]
          description: Current drive status
        origin:
          $ref: '#/components/schemas/Location'
        destination:
          $ref: '#/components/schemas/Location'
        totalDistance:
          type: number
          format: double
          description: Total distance traveled in miles
        totalDuration:
          type: integer
          description: Total duration in minutes
        settings:
          type: object
          description: Drive-specific settings
        statistics:
          type: object
          description: Drive statistics
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    DriveStartRequest:
      type: object
      required:
        - profileId
      properties:
        profileId:
          type: string
          format: uuid
        origin:
          $ref: '#/components/schemas/Location'
        destination:
          $ref: '#/components/schemas/Location'
        settings:
          type: object
    DriveUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, paused, completed, archived, cancelled]
        endTime:
          type: string
          format: date-time
        totalDistance:
          type: number
          format: double
        totalDuration:
          type: integer
        settings:
          type: object
    Location:
      type: object
      required:
        - latitude
        - longitude
        - timestamp
      properties:
        latitude:
          type: number
          format: double
          description: Latitude coordinate
        longitude:
          type: number
          format: double
          description: Longitude coordinate
        address:
          type: string
          description: Human-readable address
        city:
          type: string
        state:
          type: string
        country:
          type: string
        timestamp:
          type: string
          format: date-time

    # Voice System
...
    ConversationRequest:
      type: object
      required:
        - profileId
        - message
      properties:
        profileId:
          type: string
          format: uuid
        message:
          type: string
          description: User's voice input text
        driveId:
          type: string
          format: uuid
          description: Associated drive session
        location:
          $ref: '#/components/schemas/Location'
        context:
          type: object
          description: Conversation context
    ConversationResponse:
      type: object
      required:
        - turnId
        - role
        - text
        - timestamp
      properties:
        turnId:
          type: string
          format: uuid
        role:
          type: string
          enum: [traveler, assistant, narrator]
        text:
          type: string
          description: Response text
        audioUrl:
          type: string
          format: uri
          description: Generated audio URL
        audioDuration:
          type: number
          description: Audio duration in seconds
        timestamp:
          type: string
          format: date-time
        synopsis:
          type: string
          description: Brief summary
        metadata:
          type: object

    # POI System
...
    Error:
      type: object
      required:
        - error
        - message
        - timestamp
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
        details:
          type: object
          description: Additional error details
        requestId:
          type: string
          description: Request identifier for debugging
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
security:
  - ApiKeyAuth: []
  - BearerAuth: []
...